release version of Rest API working

+++ grails-app/services/net/hedtech/banner/virtualDomain/VirtualDomainSqlService.groovy
@@ -159,7 +159,7 @@ class VirtualDomainSqlService {
             sql.close()
         }
         println logmsg
-        return [error: errorMessage, totalCount:totalCount]
+        return [error: errorMessage, totalCount:totalCount.longValue()]
     }
 
     def update(vd, params, data) {
